+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/gtceu/WorkablePeripheralPlugin.kt
@@ -1,6 +1,7 @@
 package site.siredvin.peripheralworks.integrations.gtceu
 
 import com.gregtechceu.gtceu.api.capability.IWorkable
+import com.gregtechceu.gtceu.integration.cctweaked.peripherals.WorkablePeripheral
 import dan200.computercraft.api.lua.LuaFunction
 import dan200.computercraft.api.lua.MethodResult
 import site.siredvin.tweakium.modules.peripheral.api.IPeripheralPlugin
@@ -13,11 +14,11 @@ class WorkablePeripheralPlugin(private val workable: IWorkable) : IPeripheralPlu
         get() = TYPE
 
     @LuaFunction(mainThread = true)
-    fun getProgress(): MethodResult = MethodResult.of(workable.progress)
+    fun getProgress(): MethodResult = WorkablePeripheral.getProgress(workable)
 
     @LuaFunction(mainThread = true)
-    fun getMaxProgress(): MethodResult = MethodResult.of(workable.maxProgress)
+    fun getMaxProgress(): MethodResult = WorkablePeripheral.getMaxProgress(workable)
 
     @LuaFunction(mainThread = true)
-    fun isActive(): MethodResult = MethodResult.of(workable.isActive)
+    fun isActive(): MethodResult = WorkablePeripheral.isActive(workable)
 }